Minor adjustments
