+++ .swiftlint.yml
@@ -24,7 +24,7 @@ closure_body_length:
 file_header:
   required_pattern: |
     \/\/
-    \/\/  \w+\.\w+
+    \/\/  .*\.\w+
     \/\/  (SwiftCommonMarkTests|SwiftCommonMark|Examples)
     \/\/
     \/\/  Created by \w+ \w+ on \d+\/\d+\/\d+\.
